// src/app/api/model/getLoanProductsResponse.ts
/**
 * Apache Fineract REST API
 *
 * Contact: dev@fineract.apache.org
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GetLoanProductsInterestRateFrequencyType } from './getLoanProductsInterestRateFrequencyType';
import { GetLoanProductsRepaymentStartDateType } from './getLoanProductsRepaymentStartDateType';
import { StringEnumOptionData } from './stringEnumOptionData';
import { GetLoanProductsAmortizationType } from './getLoanProductsAmortizationType';
import { GetLoanProductsCurrency } from './getLoanProductsCurrency';
import { GetLoanProductsDaysInYearCustomStrategy } from './getLoanProductsDaysInYearCustomStrategy';
import { GetLoanProductsInterestType } from './getLoanProductsInterestType';
import { GetLoanProductsDaysInYearType } from './getLoanProductsDaysInYearType';
import { GetLoanProductsAccountingRule } from './getLoanProductsAccountingRule';
import { GetLoanProductsInterestRecalculationData } from './getLoanProductsInterestRecalculationData';
import { GetLoanProductsRepaymentFrequencyType } from './getLoanProductsRepaymentFrequencyType';
import { GetLoanProductsDaysInMonthType } from './getLoanProductsDaysInMonthType';
import { GetLoansProductsInterestCalculationPeriodType } from './getLoansProductsInterestCalculationPeriodType';


/**
 * GetLoanProductsResponse
 */
export interface GetLoanProductsResponse { 
    accountingRule?: GetLoanProductsAccountingRule;
    amortizationType?: GetLoanProductsAmortizationType;
    annualInterestRate?: number;
    buyDownFeeCalculationType?: StringEnumOptionData;
    buyDownFeeIncomeType?: StringEnumOptionData;
    buyDownFeeStrategy?: StringEnumOptionData;
    capitalizedIncomeCalculationType?: StringEnumOptionData;
    capitalizedIncomeStrategy?: StringEnumOptionData;
    capitalizedIncomeType?: StringEnumOptionData;
    chargeOffBehaviour?: StringEnumOptionData;
    currency?: GetLoanProductsCurrency;
    daysInMonthType?: GetLoanProductsDaysInMonthType;
    daysInYearCustomStrategy?: GetLoanProductsDaysInYearCustomStrategy;
    daysInYearType?: GetLoanProductsDaysInYearType;
    enableBuyDownFee?: boolean;
    enableIncomeCapitalization?: boolean;
    endDate?: string;
    fixedLength?: number;
    fixedPrincipalPercentagePerInstallment?: number;
    id?: number;
    includeInBorrowerCycle?: boolean;
    interestCalculationPeriodType?: GetLoansProductsInterestCalculationPeriodType;
    interestRateFrequencyType?: GetLoanProductsInterestRateFrequencyType;
    interestRatePerPeriod?: number;
    interestRateVariationsForBorrowerCycle?: Array<number>;
    interestRecalculationData?: GetLoanProductsInterestRecalculationData;
    interestRecognitionOnDisbursementDate?: boolean;
    interestType?: GetLoanProductsInterestType;
    isInterestRecalculationEnabled?: boolean;
    maxNumberOfRepayments?: number;
    maxPrincipal?: number;
    merchantBuyDownFee?: boolean;
    minNumberOfRepayments?: number;
    minPrincipal?: number;
    name?: string;
    numberOfRepaymentVariationsForBorrowerCycle?: Array<number>;
    numberOfRepayments?: number;
    principal?: number;
    principalThresholdForLastInstalment?: number;
    principalVariationsForBorrowerCycle?: Array<number>;
    repaymentEvery?: number;
    repaymentFrequencyType?: GetLoanProductsRepaymentFrequencyType;
    repaymentStartDateType?: GetLoanProductsRepaymentStartDateType;
    shortName?: string;
    startDate?: string;
    status?: string;
    supportedInterestRefundTypes?: Array<StringEnumOptionData>;
    transactionProcessingStrategy?: string;
    transactionProcessingStrategyName?: string;
    useBorrowerCycle?: boolean;
}